draw_service/pyproject.toml
[project]
name = "draw_service"
version = "0.1.0"
description = ""
authors = [{ name="Moshe", email="barboym@gmail.com" }]
dependencies = [
    "fastapi[standard]",
    "Pillow",
    "uvicorn",
    "scipy",
    "psycopg2-binary",
    "SQLAlchemy>=2.0",
    "python-dotenv",
    "pytest",
    "llama-index-llms-anthropic",
    "llama-index",
]
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["chatdraw"]
